+++ src/vector/arithmetic/vector_add_d.cpp
@@ -8,7 +8,7 @@ using namespace ddavx_core;
 
 void d_real_vector::add(dd_real_vector& vec1, dd_real_vector& vec2)
 {
-	if((size_t)size() != (size_t)vec1.size() || (size_t)size() != (size_t)vec2.size()){
+	if(size() != vec1.size() || size() != vec2.size()){
 		std::cerr << "error bad vector size" << std::endl;
 		assert(1);
 	}
@@ -39,7 +39,7 @@ void d_real_vector::add(dd_real_vector& vec1, dd_real_vector& vec2)
 
 void d_real_vector::add(d_real_vector& vec1, dd_real_vector& vec2)
 {
-	if((size_t)size() != (size_t)vec1.size() || (size_t)size() != (size_t)vec2.size()){
+	if(size() != vec1.size() || size() != vec2.size()){
 		std::cerr << "error bad vector size" << std::endl;
 		assert(1);
 	}
@@ -76,13 +76,13 @@ void d_real_vector::add(dd_real_vector& vec1, d_real_vector& vec2)
 
 void d_real_vector::add(d_real_vector& vec1, d_real_vector& vec2)
 {
-	if((size_t)size() != (size_t)vec1.size() || (size_t)size() != (size_t)vec2.size()){
+	if(size() != vec1.size() || size() != vec2.size()){
 		std::cerr << "error bad vector size" << std::endl;
 		assert(1);
 	}
 
 #pragma omp parallel for
-	for(size_t i = 0; i < (size_t)size(); i++){
+	for(size_t i = 0; i < size(); i++){
 		data()[i] = vec1.data()[i] + vec2.data()[i];
 
 	}
