+++ src/vector/arithmetic/vector_add_dd.cpp
@@ -8,7 +8,7 @@ using namespace ddavx_core;
 
 void dd_real_vector::add(dd_real_vector& vec1, dd_real_vector& vec2)
 {
-	if((size_t)size() != (size_t)vec1.size() || (size_t)size() != (size_t)vec2.size()){
+	if(size() != vec1.size() || size() != vec2.size()){
 		std::cerr << "error bad vector size" << std::endl;
 		assert(1);
 	}
@@ -41,7 +41,7 @@ void dd_real_vector::add(dd_real_vector& vec1, dd_real_vector& vec2)
 
 void dd_real_vector::add(d_real_vector& vec1, dd_real_vector& vec2)
 {
-	if((size_t)size() != (size_t)vec1.size() || (size_t)size() != (size_t)vec2.size()){
+	if(size() != vec1.size() || size() != vec2.size()){
 		std::cerr << "error bad vector size" << std::endl;
 		assert(1);
 	}
@@ -80,7 +80,7 @@ void dd_real_vector::add(dd_real_vector& vec1, d_real_vector& vec2)
 
 void dd_real_vector::add(d_real_vector& vec1, d_real_vector& vec2)
 {
-	if((size_t)size() != (size_t)vec1.size() || (size_t)size() != (size_t)vec2.size()){
+	if(size() != vec1.size() || size() != vec2.size()){
 		std::cerr << "error bad vector size" << std::endl;
 		assert(1);
 	}
