+++ src/vector/blas/xpay.cpp
@@ -5,7 +5,7 @@ using namespace ddavx_core;
 namespace dd_avx{
 //alpha = DD ///////////////////////////////////////////
 	void xpay(const dd_real& alpha, const dd_real_vector& x, dd_real_vector& y){
-		if((size_t)x.size() != (size_t)y.size()){
+		if(x.size() != y.size()){
 			std::cerr << "error bad vector size" << std::endl;
 			assert(1);
 		}
@@ -14,7 +14,7 @@ namespace dd_avx{
 #pragma omp parallel private(regs)
 		{
 			size_t i=0, is=0, ie=0;
-			get_isie((size_t)y.size(), is, ie);
+			get_isie(y.size(), is, ie);
 			AVXreg alpha_hi = broadcast(alpha.x[0]);
 			AVXreg alpha_lo = broadcast(alpha.x[1]);
 			for(i = is; i < ie - SIMD_Length - 1; i += SIMD_Length){
@@ -37,7 +37,7 @@ namespace dd_avx{
 	}
 
 	void xpay(const dd_real& alpha, const d_real_vector& x, dd_real_vector& y){
-		if((size_t)x.size() != (size_t)y.size()){
+		if(x.size() != y.size()){
 			std::cerr << "error bad vector size" << std::endl;
 			assert(1);
 		}
@@ -46,7 +46,7 @@ namespace dd_avx{
 #pragma omp parallel private(regs)
 		{
 			size_t i=0, is=0, ie=0;
-			get_isie((size_t)y.size(), is, ie);
+			get_isie(y.size(), is, ie);
 			AVXreg alpha_hi = broadcast(alpha.x[0]);
 			AVXreg alpha_lo = broadcast(alpha.x[1]);
 			for(i = is; i < ie - SIMD_Length - 1; i += SIMD_Length){
@@ -69,7 +69,7 @@ namespace dd_avx{
 	}
 
 	void xpay(const dd_real& alpha, const dd_real_vector& x, d_real_vector& y){
-		if((size_t)x.size() != (size_t)y.size()){
+		if(x.size() != y.size()){
 			std::cerr << "error bad vector size" << std::endl;
 			assert(1);
 		}
@@ -78,7 +78,7 @@ namespace dd_avx{
 #pragma omp parallel private(regs)
 		{
 			size_t i=0, is=0, ie=0;
-			get_isie((size_t)y.size(), is, ie);
+			get_isie(y.size(), is, ie);
 			AVXreg alpha_hi = broadcast(alpha.x[0]);
 			AVXreg alpha_lo = broadcast(alpha.x[1]);
 			for(i = is; i < ie - SIMD_Length - 1; i += SIMD_Length){
@@ -100,7 +100,7 @@ namespace dd_avx{
 	}
 
 	void xpay(const dd_real& alpha, const d_real_vector& x, d_real_vector& y){
-		if((size_t)x.size() != (size_t)y.size()){
+		if(x.size() != y.size()){
 			std::cerr << "error bad vector size" << std::endl;
 			assert(1);
 		}
@@ -109,7 +109,7 @@ namespace dd_avx{
 #pragma omp parallel private(regs)
 		{
 			size_t i=0, is=0, ie=0;
-			get_isie((size_t)y.size(), is, ie);
+			get_isie(y.size(), is, ie);
 			AVXreg alpha_hi = broadcast(alpha.x[0]);
 			AVXreg alpha_lo = broadcast(alpha.x[1]);
 			for(i = is; i < ie - SIMD_Length - 1; i += SIMD_Length){
@@ -132,7 +132,7 @@ namespace dd_avx{
 
 //alpha = D ///////////////////////////////////////////
 	void xpay(const d_real& alpha, const dd_real_vector& x, dd_real_vector& y){
-		if((size_t)x.size() != (size_t)y.size()){
+		if(x.size() != y.size()){
 			std::cerr << "error bad vector size" << std::endl;
 			assert(1);
 		}
@@ -141,7 +141,7 @@ namespace dd_avx{
 #pragma omp parallel private(regs)
 		{
 			size_t i=0, is=0, ie=0;
-			get_isie((size_t)y.size(), is, ie);
+			get_isie(y.size(), is, ie);
 			AVXreg alpha_hi = broadcast(alpha);
 			AVXreg alpha_lo = regs.zeros;
 			for(i = is; i < ie - SIMD_Length - 1; i += SIMD_Length){
@@ -164,7 +164,7 @@ namespace dd_avx{
 	}
 
 	void xpay(const d_real& alpha, const d_real_vector& x, dd_real_vector& y){
-		if((size_t)x.size() != (size_t)y.size()){
+		if(x.size() != y.size()){
 			std::cerr << "error bad vector size" << std::endl;
 			assert(1);
 		}
@@ -173,7 +173,7 @@ namespace dd_avx{
 #pragma omp parallel private(regs)
 		{
 			size_t i=0, is=0, ie=0;
-			get_isie((size_t)y.size(), is, ie);
+			get_isie(y.size(), is, ie);
 			AVXreg alpha_hi = broadcast(alpha);
 			AVXreg alpha_lo = regs.zeros;
 			for(i = is; i < ie - SIMD_Length - 1; i += SIMD_Length){
@@ -196,7 +196,7 @@ namespace dd_avx{
 	}
 
 	void xpay(const d_real& alpha, const dd_real_vector& x, d_real_vector& y){
-		if((size_t)x.size() != (size_t)y.size()){
+		if(x.size() != y.size()){
 			std::cerr << "error bad vector size" << std::endl;
 			assert(1);
 		}
@@ -205,7 +205,7 @@ namespace dd_avx{
 #pragma omp parallel private(regs)
 		{
 			size_t i=0, is=0, ie=0;
-			get_isie((size_t)y.size(), is, ie);
+			get_isie(y.size(), is, ie);
 			AVXreg alpha_hi = broadcast(alpha);
 			AVXreg alpha_lo = regs.zeros;
 			for(i = is; i < ie - SIMD_Length - 1; i += SIMD_Length){
@@ -227,7 +227,7 @@ namespace dd_avx{
 	}
 
 	void xpay(const d_real& alpha, const d_real_vector& x, d_real_vector& y){
-		if((size_t)x.size() != (size_t)y.size()){
+		if(x.size() != y.size()){
 			std::cerr << "error bad vector size" << std::endl;
 			assert(1);
 		}
@@ -236,7 +236,7 @@ namespace dd_avx{
 #pragma omp parallel private(regs)
 		{
 			size_t i=0, is=0, ie=0;
-			get_isie((size_t)y.size(), is, ie);
+			get_isie(y.size(), is, ie);
 			AVXreg alpha_hi = broadcast(alpha);
 			AVXreg alpha_lo = regs.zeros;
 			for(i = is; i < ie - SIMD_Length - 1; i += SIMD_Length){
